model_optimization/python_scripts/requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
pyyaml==6.0.2
